Railway Reservation System
The Railway Reservation System facilitates the passengers to enquire about the trains available on the basis of source and destination, booking of tickets, generating cost of order, and enquire about the status of the booked ticket using JAVA. It is the computerized system of reserving the seats of train seats in advance. It is mainly used for long routes.