Airline system
The project is an effort to provide a one-step solution to the customer with a system that can provide them with information regarding availability of flight tickets. The system provides a way to make ticket bookings for the customers. The availability of the flight tickets can only be updated by authorized airlines personnel, and hence it becomes easy for the Customer to select their ticket from the available flights. We have implemented Agile Methodology while designing the system.