City Nursing Home Website
The key idea of this project was enabling the users to book an appointment with a specialist (cardiologist, physician, OBS/Gynecologist) doctor at the hospital. The database was procured from the hospital’s website and managed by incorporating phpMyAdmin. In case an appointment slot was booked for a particular doctor by a patient, then that booking slot was made unavailable for all other users to preserve consistency. Further, we equipped the site to maintain accounts for each patient. The administrators could then view the medical history and appointment slots, as we had included features like time, history, doctor, etc. to link the client, server, and interface files. This project was pivotal to ingraining technical expertise in HTML5, CSS, and JavaScript.